backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-filter==24.3
django-cors-headers==4.6.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
openpyxl==3.1.5
reportlab==4.2.5
